This commit adds configure script checking for a number of functions in libm:  - sinf, cosf, atanf, atan2f, ...
Previously, in x_arithmetic.c, these functions were redefined for MSW / Apple.
Now, they will only be redefined if the configure script does not find them.
The results of the checks for these functions are placed in the config.h header
file.  The config.h.in file was generated using autoheader.

This commit also introduces a library of m4 macros that I use in another
project to make adding these checks easier, as well as other things in the
future.  I left the macro prefixes as AST_ to make it easier to keep this file
in sync with the original version.

Important Note:
 - With the introduction of acinclude.m4, you have to run aclocal before
   running autoconf.  It wouldn't hurt to make running autoheader a habit
   as well.  We should probably just add a bootstrap/autogen script that does
   all of this for you.