This fixes bug 1572290 on Windows, and should help on the other platforms too. I need to test the others now.
The problem was that the Tcl internal character encoding was just being sent straight to the pd process without being converted to the current system encoding.  I added Tcl_UtfToExternalDString() as an easy way to convert from the Tcl encoding to the current system encoding and it seems to work fine now.